test: move + update
